refactor: explicitely added 'use' to the beginning of all hooks file names
