+++ src/components/DownloadButton/DownloadButton.jsx
@@ -4,7 +4,7 @@ import { Tooltip } from 'flowbite-react';
 
 import ModalDownloadRewiews from './ModalDownloadRewiews';
 import { buildFullApiUrl, sendDownloadApiRequest, sendSaveQIdApiRequest } from '../../lib/istexApi';
-import useResetForm from '@/hooks/resetForm';
+import useResetForm from '@/hooks/useResetForm';
 import { useEventEmitterContext } from '@/contexts/EventEmitterContext';
 import { useHistoryContext } from '@/contexts/HistoryContext';
