+++ src/components/QueryInput/QueryInput.jsx
@@ -20,8 +20,8 @@ import {
 } from '../../lib/istexApi';
 import { queryModes, istexApiConfig } from '../../config';
 import { setQueryString, setQId } from '../../store/istexApiSlice';
-import useFocus from '@/hooks/focus';
-import useResetForm from '@/hooks/resetForm';
+import useFocus from '@/hooks/useFocus';
+import useResetForm from '@/hooks/useResetForm';
 import { useUrlSearchParamsContext } from '@/contexts/UrlSearchParamsContext';
 import { useEventEmitterContext } from '@/contexts/EventEmitterContext';
